benchmarking: split the glutton package and share its route constants
glutton.go arrived from the previous commit as a single 640-line file, and the
routes it serves were spelled out twice: once in the mux, once in the fake that
stands in for it in boomer's tests.

Split it by concern into server.go, service.go, metrics.go, and gossip.go,
leaving glutton.go with the package doc and the constants describing the
workload's contract. The instrument block moves out of New into
Service.initMetrics.

Add Name, ModeGRPC/ModeHTTP, and one constant per route. The fake re-exports
the route constants instead of declaring its own, so a renamed path cannot
leave the stand-in answering something the actor does not. Name replaces the
four "glutton" literals across the tracer scope, meter scope, and the binary's
tracing and metrics init. Instrument names stay written out in full, since
hack/verify/metrics.sh greps for them.

Boomer keeps its own private route constants: it is the client, and pointing it
at the server's implementation package would link the gRPC server and otel
instruments into a binary that runs neither.

The tests split the same way, and gain coverage for Handler's mode selection
and the /writeram route.
